requirements.txt
flask
flask-cors
google-generativeai
pinecone-client
python-dotenv